Revert "Added UV and texture support into the engine"
This reverts commit 6fc92af502e1d3399f13b3cb2597efc75f41636e.

+++ Source/Object/Mesh/2D/NRE_Mesh2D.cpp
@@ -33,7 +33,7 @@
 
                 add(0, vData, 18);
 
-                GLfloat uvData[] = {
+                GLshort uvData[] = {
                     0, 0,
                     1, 0,
                     1, 1,
